fix: resolve clippy warnings, unreachable code, and compilation failures
- Fixed type mismatch in src/net/quic_core/mod.rs related to packet numbers
- Restored missing Duration import in src/net/unix/stream.rs
- Corrected dangling TaskComplete match arms in src/obligation/* files
- Addressed pervasive clippy warnings across channel, io, stream, and conformance tests
- Fixed type inference error in mpsc.rs projections HashMap
- Updated test IDs to new_ephemeral in resource_cleanup_verifier

+++ src/net/quic_core/mod.rs
@@ -2085,12 +2085,11 @@ mod tests {
 
                 // Reconstruct using the original as "largest_pn"
                 let reconstructed =
-                    decode_packet_number_reconstruct(truncated, width, packet_number as u64)
-                        .unwrap();
+                    decode_packet_number_reconstruct(truncated, width, packet_number).unwrap();
 
                 // Should recover the original packet number
                 assert_eq!(
-                    reconstructed, packet_number as u64,
+                    reconstructed, packet_number,
                     "Round-trip reconstruction failed for 0x{packet_number:x} width {width}"
                 );
             }
